admission: move some grant coordinators to their own source file
Specifically ElasticCPUGrantCoordinator and StoreGrantCoordinators are
moved. And there is some tiny renaming and addition of commentary.

This is a precursor to simplifying how StoreGrantCoordinators is
implemented, since it doesn't use most of the functionality in
GrantCoordinator (as it doesn't need to manange multiple granters).

Epic: none

Release note: None
